+++ sparkmagic/sparkmagic/auth/basic.py
@@ -18,7 +18,7 @@ class Basic(HTTPBasicAuth, Authenticator):
             is created from parsing %spark magic command.
         """
         if parsed_attributes is not None:
-            if parsed_attributes.user is '' or parsed_attributes.password is '':
+            if parsed_attributes.user == '' or parsed_attributes.password == '':
                 new_exc = BadUserDataException("Need to supply username and password arguments for "\
                     "Basic Access Authentication. (e.g. -a username -p password).")
                 raise new_exc
@@ -47,7 +47,7 @@ class Basic(HTTPBasicAuth, Authenticator):
             width=widget_width
         )
 
-        self.password_widget = ipywidget_factory.get_text(
+        self.password_widget = ipywidget_factory.get_password(
             description='Password:',
             value=self.password,
             width=widget_width