+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Features
 Added one internal magic to enable retry of session creation. 
 
+* New `%%pretty` magic for pretty printing a dataframe as an HTML table. Thanks @hegary 
+* Update Endpoint widget to shield passwords when entering them in the ipywidget. Thanks @J0rg3M3nd3z @jodom961
+
 ## 0.18.0
 
 ### Updates