+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## NEXT RELEASE
 
+## 0.19.0
+
+### Features
+
+* Added one internal magic to enable retry of session creation. Thanks @edwardps
+* New `%%pretty` magic for pretty printing a dataframe as an HTML table. Thanks @hegary 
+* Update Endpoint widget to shield passwords when entering them in the ipywidget. Thanks @J0rg3M3nd3z @jodom961
+
 ## 0.18.0
 
 ### Updates